fix(wasm): Fix unwrap error on Wasm platforms caused by UInt::MAX conversions (#5240)
UInt::MAX.try_into().unwrap() was causing errors on Wasm platforms, due
to the result being unrepresentable.

This `unwrap_or` was also always being calculated regardless, so I think
using `usize::MAX` is preferable on all platforms.

+++ bindings/matrix-sdk-ffi/src/timeline/msg_like.rs
@@ -241,7 +241,7 @@ pub struct PollAnswer {
 #[derive(Clone, uniffi::Object)]
 pub struct ThreadSummary {
     pub latest_event: EmbeddedEventDetails,
-    pub num_replies: usize,
+    pub num_replies: u32,
 }
 
 #[matrix_sdk_ffi_macros::export]
